+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.1-prerelease] - 2023-06-26
+
+-Made GameTags into a static class for runtime and editor usage. No more Prefab required.
+-Changed the default settings location to GameTags/Resources so that they may be found at runtime.
+-Fixed a bug where GameTagContainer would always return invalid at editor rime.
+
 ## [0.1.1] - 2023-06-07
 
 ### Added